Revert "Melhorias, correções e/ou pequenos ajustes no código."
This reverts commit 7a8749efdbf53be86d00e1a3719d8db86395a918.
