+++ .gitignore
@@ -4,4 +4,3 @@ npm-debug.log
 .env
 .DS_Store
 ffmpeg.exe
-commit.txt
